add semantic errors

+++ src/semantic/errors.rs
@@ -3,5 +3,5 @@ use thiserror::Error;
 #[derive(Error, Debug)]
 pub enum SemanticError {
     #[error("Expected {expected} arguments, found {found}")]
-    ArgumentCountMismatch { expected: i32, found: i32 },
+    ArgumentCountMismatch { expected: usize, found: usize },
 }
